pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "documentor"
version = "0.2.0"
description = "AI-powered PDF document classification and organization using vision LLMs"
authors = [
    { name = "Tiago Silva", email = "eng.tiago.silva@gmail.com" }
]
dependencies = [
    "tqdm",
    "python-dotenv",
    "openai",
    "PyMuPDF",
    "pandas",
    "pydantic",
    "openpyxl",
    "pillow"
]
requires-python = ">=3.10"

[project.scripts]
documentor = "main:main"
check-hash = "scripts.check_hash:main"

[tool.hatch.build.targets.wheel]
include = [
    "main.py",
    "documentor/**",
    "config/**",
    "scripts/**"
]
